HashCode Docstrings

Documentation for HashCode.

HashCode.CarPath

HashCode.CityMap

HashCode.DFS

HashCode.addCarMove

HashCode.alreadyVisited

HashCode.custom_walk

HashCode.djk

HashCode.findMaximumEndpoints

HashCode.getCarLocation

HashCode.getCarLocation

HashCode.getJuncOptions

HashCode.getMinUnseen

HashCode.getPathTime

HashCode.getPathToNode

HashCode.getPosJuncs

HashCode.haveTraversed

HashCode.isFeasible

HashCode.laterRandomWalk

HashCode.CarPath — Type

CarPath

A structure representing the path and travel information of a car in the city.

Fields:

carNum (Int): The unique identifier for the car.

path (Vector{Int}): Vector representing the path of the car, starting with an initial junction.

timeTraveled (Int): The total time traveled by the car.

Constructor:

CarPath(carNum::Int): Creates a new CarPath object with the specified car number.

carNum: The unique identifier for the car.

HashCode.CityMap — Type

CityMap

A structure representing the map of a city, including information about nodes (junctions), edges (streets), cars, and visited junctions.

Fields:

nodes (Vector{Junction}): Vector of junctions in the city.

edges (Vector{Street}): Vector of streets connecting junctions.

cars (Vector{CarPath}): Vector of car paths representing the movement of cars.

visitedJuncs (Vector{Int}): Vector of visited junctions.

Constructors:

CityMap(junctions::Vector{Junction}, streets::Vector{Street}): Creates a new CityMap with the given junctions and streets.

junctions: Vector of Junction objects.

streets: Vector of Street objects.

Methods:

newFunc(): Returns the string "hi".

HashCode.DFS — Method

exploreEdgesDFS(city::City, curJunc::Int, curTime::Int, maxTime::Int)

Runs a Depth-First Search (DFS) to find all possible edges that are reachable within the specified time limit. This function serves as a sensible upper bound for the best possible performance with 8 cars, as it calculates the sum of the distances of every reachable node.

Parameters:

city (City): The city object containing information about nodes and edges.

curJunc (Int): The current junction from which to find possible next junctions.

curTime (Int): The total amount of time that has elapsed up until this recursive call.

maxTime (Int): The maximum time limit for the cars to travel.

Returns:

Nothing: Modifies a global variable distances, which is a list of distances of reachable edges. The sum of these distances represents the total distance and serves as an upper bound.

HashCode.addCarMove — Method

addCarMove(cityMap::CityMap, carNum::Int, juncNum::Int, timeAdded::Int)

Updates the city map to reflect a car's movement by adding a new junction to its path and updating its total travel time.

Parameters:

cityMap (CityMap): The map of the city containing information about cars and visited junctions.

carNum (Int): The index of the car making the move.

juncNum (Int): The junction number to be added to the car's path.

timeAdded (Int): The time added to the car's total travel time.

Returns:

Nothing

HashCode.alreadyVisited — Method

alreadyVisited(cityMap::CityMap, juncNum::Int)

Checks if a junction has already been visited in the city map.

Parameters:

cityMap (CityMap): The map of the city containing information about visited junctions.

juncNum (Int): The junction number to check for.

Returns:

Bool: True if the junction has already been visited, false otherwise.

HashCode.custom_walk — Method

custom_walk(city::City, numTrials::Int)

Performs a custom walk to generate itineraries for vehicles in a given city.

Parameters:

rng (AbstractRNG): The random number generator.

city (City): The city information containing details such as total duration, number of cars, starting junction, and streets.

Returns:

TrialResults: A list of the results from running numTrials trials

HashCode.djk — Method

djk(city::City, startJunc::Int)

Djikstra's algorithm to identify the shortest path from the start point to every other node in the graph. Uses standard Djikstra's implementation

Parameters:

city (City): The city object containing information about nodes and edges.

startJunc (junction): junction at which to start the search

Returns:

distances: returns vector that contains the minimum distance from the start node to each of the other nodes in the graph

previous: returns a vector that, for each node, has the previous node that would go in front of it in the shortest path this is done in standard djikstra formulation.

HashCode.findMaximumEndpoints — Method

findMaximumEndpoints()

Scans through each of the junctions in the city graphs and identitifies the furtherest point in each of the 8 cardinal directions

Parameters: None

Returns:

List of the most extreme junctions in each of the cardinal directions that will serve as the target points for each of the 8 cars

HashCode.getCarLocation — Method

getCarLocation(car::CarPath)

Gets the current location of a car based on its path.

Parameters:

car (CarPath): The car path object containing information about the car's path.

Returns:

Any: The current location of the car.

HashCode.getCarLocation — Method

getCarLocation(cityMap::CityMap, carNum::Int)

Gets the current location of a specified car in the city map.

Parameters:

cityMap (CityMap): The map of the city containing information about cars.

carNum (Int): The index of the car for which to retrieve the location.

Returns:

Any: The current location of the specified car.

HashCode.getJuncOptions — Method

getPosJuncs(city::City, curJunc::Int)

Finds the possible junctions and their durations reachable from the given current junction in a city map.

Parameters:

city (City): The city object containing information about nodes and edges.

curJunc (Int): The current junction from which to find possible next junctions.

Returns:

Vector{Tuple{Int, Int}}: A vector of tuples representing possible next junctions and their corresponding distances and durations.

HashCode.getMinUnseen — Method

getMinUnseen(unseenNodes, distances)

Helper function for Djikstra's algorithm. Given a set of unseen nodes and the djikstra's distances vector, identifies the minimum distance node to use next in the search.

Parameters:

unseenNodes: set of unseen/unvisited nodes that we must choose from

distances: djikstra style distances vector that has the current mininum distance to each node

Returns:

minNode: node that has the minimum current distance amoung unvisited nodes

HashCode.getPathTime — Method

getPathTime(path :: Vector{Int},city :: City)

Given a path, gets the total time that a car has spent on that path in order to understand how much time the car has left

Parameters:

path: list of nodes that form the path that the car has already traveled

city: standard city object

Returns:

totalTime: The total time that the car has already spent traveling

HashCode.getPathToNode — Method

getPathToNode(endNode::Int, previous)

Given a node and the previous set, reconstructs the shortest path from the starting point to that end node

Parameters:

endNode: node to create the path to

previous: a vector that, for each node, has the previous node that would go in front of it in the shortest path

this is done in standard djikstra formulation.

Returns:

nodeList: list of nodes that form the shortest path from the startpoint (4517) to the node

HashCode.getPosJuncs — Method

getPosJuncs(cityMap::CityMap, curJunc::Int)

Finds the possible junctions and their durations reachable from the given current junction in a city map.

Parameters:

cityMap (CityMap): The map of the city containing information about edges.

curJunc (Int): The current junction from which to find possible next junctions.

Returns:

Vector{Tuple{Int, Int}}: A vector of tuples representing possible next junctions and their corresponding durations.

HashCode.haveTraversed — Method

haveTraversed(curJunc::Int, posJunc)

Checks if a connection between two junctions has been traversed.

Parameters:

curJunc (Int): The current junction.

posJunc (Tuple{Int, Int}): A tuple representing a possible next junction and its duration.

Returns:

Bool: True if the connection has not been traversed, false otherwise.

HashCode.isFeasible — Method

isFeasible(solution::HashCode2014.Solution, city::HashCode2014.City)

Checks the feasibility of a given solution within the context of the provided city.

Parameters:

solution (HashCode2014.Solution): The solution to be checked for feasibility.

city (HashCode2014.City): The city object containing information about nodes and edges.

Returns:

Bool: True if the solution is feasible within the given city, False otherwise.

HashCode.laterRandomWalk — Method

laterRandomWalk(previous :: Vector{Int}, city:: Int)

Given the results of the djikstra algorithm, travels the cars to the endpoints and then random walks them around the area to cover the most distance possible

Parameters:

previous: a vector that, for each node, has the previous node that would go in front of it in the shortest path

this is done in standard djikstra formulation.

city: standard city object containing information about the junctions and streets

Returns:

itineraries: set of num_cars itineraries that represent the suggest paths for each car to be used in the solution